# payment-service/src/main/java/payment_service/PaymentServiceDocSwagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Payment Service API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8083",
      "description": "Local Payment Service"
    }
  ],

  "paths": {

    "/payments": {
      "post": {
        "tags": ["Payment"],
        "summary": "Create payment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PaymentRequestDTO" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaymentResponseDTO" }
              }
            }
          },
          "400": { "description": "Invalid request" },
          "404": { "description": "User or delivery not found" }
        }
      }
    },

    "/payments/{paymentId}/start": {
      "post": {
        "tags": ["Payment"],
        "summary": "Start payment",
        "parameters": [
          { "name": "paymentId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Payment started" },
          "404": { "description": "Payment not found" }
        }
      }
    },

    "/payments/{paymentId}/cancel": {
      "post": {
        "tags": ["Payment"],
        "summary": "Cancel payment",
        "parameters": [
          { "name": "paymentId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Canceled" },
          "404": { "description": "Payment not found" }
        }
      }
    },

    "/payments/{paymentId}": {
      "get": {
        "tags": ["Payment"],
        "summary": "Get payment state",
        "parameters": [
          { "name": "paymentId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "State",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "enum": ["REQUESTED", "PROCESSING", "SUCCESS", "FAILED", "CANCELED", "REFUND_REQUESTED", "REFUNDING", "REFUNDED"]
                }
              }
            }
          },
          "404": { "description": "Payment not found" }
        }
      }
    },

    "/payments/{paymentId}/refund": {
      "post": {
        "tags": ["Payment"],
        "summary": "Request and start refund",
        "parameters": [
          { "name": "paymentId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Refund started" },
          "404": { "description": "Payment not found" },
          "409": { "description": "Invalid state" }
        }
      }
    }
  },

  "components": {
    "schemas": {

      "PaymentRequestDTO": {
        "type": "object",
        "properties": {
          "deliveryId": { "type": "string" },
          "userId": { "type": "string" },
          "amount": {
            "type": "object",
            "properties": {
              "value": { "type": "number", "format": "double" },
              "currency": { "type": "string" }
            }
          }
        }
      },

      "PaymentResponseDTO": {
        "type": "object",
        "properties": {
          "paymentId": { "type": "string" },
          "state": { "type": "string" },
          "deliveryId": { "type": "string" },
          "userId": { "type": "string" },
          "amount": {
            "type": "object",
            "properties": {
              "value": { "type": "number" },
              "currency": { "type": "string" }
            }
          }
        }
      }
    }
  }
}